Support a task created with qiskit_experiment
The JSON encoder and decoder of qiskit and qiskit_experiment are
different. Some objects can be encoded by one of encoder but the other,
and some objects can be encoded by both encoders, but the output of
two encoders are different. Additionally, when decoding the JSON files,
python pass the JSON object to object_hook() function from the bottom
of the object, so it's not possible to delegate the decoding process
depending on the top level of the JSON object.

As an workaround to correctly encode and decode the objects from qiskit
and qiskit experiments, we encode the results from the two libraries
into JSON strings and attach the strings to the corresponding wrapper
classes. And then we use the normal encoding process to encode the
wrapper classes. The disadvantage of this solution is that we encode
and decode the results from the libraries twice, and there are many
escaped strings in the final JSON files.