Support Experiments initialized with real backends
The JSON encoder and decoder provided by Qiskit Experiments cannot
properly encode and decode the instances of Experiment initialized with
real backends. To cache and recover the tasks of experiments, we have
to delete the backend entry before encoding the task. The issue below
provides an workaround to add the backend back after decoding.

https://github.com/qiskit-community/qiskit-experiments/issues/1508