Do not restrict the rotation angle less than pi
The benifit to restrict the rotation angle less than pi is that the line
of the rotation angle will reflect at pi if the pulse is over rotating,
so we can easily see whether the over rotation happens from the diagram.
However, we cannot get the real rotation angle if restricting the value
of the rotation angle.