Adjust tolerance of decomposing unitary matrices
Even the tolerance is set to 1e-12 while evolving the unitary, the error
of unitary decomposition is still sometimes higher than the accepted
tolerance. Lowering the tolerance for the ODE solver to solve this
problem will increase the computation time.

To assure the accuracy of decomposition of unitary matrices, the
direction of the rotation axis should be accurate as possible. When the
error is relatively smaller than the 1e-6 of lenth of the axis, the
error of the angle of the axis is roughly 5e-5 degrees. This is accepted
in my research, so only raise the expection when the error is relatively
larger than the 1e-6 of length of the axis.