Use default ODE solver to do unitary evolution
After benchmarking with the default ODE solver (DOP853) and
`jax_odeint`, we found `DOP853` is much faster than `jax_odeint` with
this extent of tolerance.

The behaviours of JAX kind ODE solvers are different from other solvers.
JAX kind solvers will not really solve the ODE until we query the
result. To get the fair benchmark result from the timer, we explicilty
convert the ODE result to a list before the end of the timer.