Improve unitary decomposition algorithm
Introduce a two-stage method to remove the global phase. The new method
prevents the errors above the tolerance by estimating the global phase
with the average phase from the four complex numbers containing in the
input unitary operator.

The largest error occuring with the new method is about 1e-13, which is
much smaller than the assigned tolerance values.